Add hash method
Start with an empty tag( hashCode). For each letter in the key:   hashCode = (primeNumber * hashCode + key.charCodeAt(i)) % size

We want to make sure that our tag fits into the size of hash table. So, we use the modulo operator (%) to find the remainder when dividing the tag by
the size of the box. This ensures that our tag stays within the range of possible indices in our hash table.